update start/end date calculations
Because of the changes to Date#parse in the 1.9
series of interpreters we can't really use it in
this manner. The dates we get are valid but they
aren't the dates we want. Date#civil doesn't have
any guessing involved so we'll get the same date
regardless of Ruby version.

Signed-off-by: t.pickett66 <t.pickett66@gmail.com>